# DRILL-7450: Improve performance for ANALYZE command