… docs/dev/MetastoreAnalyze.md
… exec/java-exec/src/main/java/org/apache/drill/exec/metastore/analyze/MetadataAggregateContext.java
… exec/java-exec/src/main/java/org/apache/drill/exec/physical/config/HashToMergeExchange.java
… exec/java-exec/src/main/java/org/apache/drill/exec/physical/config/MetadataAggPOP.java
… exec/java-exec/src/main/java/org/apache/drill/exec/physical/config/MetadataHashAggPOP.java
… exec/java-exec/src/main/java/org/apache/drill/exec/physical/config/MetadataStreamAggPOP.java
… exec/java-exec/src/main/java/org/apache/drill/exec/physical/impl/aggregate/HashAggBatch.java
… exec/java-exec/src/main/java/org/apache/drill/exec/physical/impl/aggregate/HashAggTemplate.java
… exec/java-exec/src/main/java/org/apache/drill/exec/physical/impl/aggregate/StreamingAggBatch.java
… exec/java-exec/src/main/java/org/apache/drill/exec/physical/impl/flatten/FlattenRecordBatch.java
… exec/java-exec/src/main/java/org/apache/drill/exec/physical/impl/metadata/MetadataAggBatch.java
… exec/java-exec/src/main/java/org/apache/drill/exec/physical/impl/metadata/MetadataAggregateHelper.java
… exec/java-exec/src/main/java/org/apache/drill/exec/physical/impl/metadata/MetadataControllerBatch.java
… exec/java-exec/src/main/java/org/apache/drill/exec/physical/impl/metadata/MetadataHandlerBatch.java
… exec/java-exec/src/main/java/org/apache/drill/exec/physical/impl/metadata/MetadataHashAggBatch.java
… exec/java-exec/src/main/java/org/apache/drill/exec/physical/impl/metadata/MetadataHashAggBatchCreator.java
… exec/java-exec/src/main/java/org/apache/drill/exec/physical/impl/metadata/MetadataStreamAggBatch.java
… exec/java-exec/src/main/java/org/apache/drill/exec/physical/impl/metadata/MetadataStreamAggBatchCreator.java
… exec/java-exec/src/main/java/org/apache/drill/exec/physical/impl/project/ProjectRecordBatch.java
… exec/java-exec/src/main/java/org/apache/drill/exec/physical/impl/validate/BatchValidator.java
… exec/java-exec/src/main/java/org/apache/drill/exec/physical/resultSet/model/single/BaseReaderBuilder.java
… exec/java-exec/src/main/java/org/apache/drill/exec/physical/rowSet/RowSetFormatter.java
… exec/java-exec/src/main/java/org/apache/drill/exec/planner/PlannerPhase.java
… exec/java-exec/src/main/java/org/apache/drill/exec/planner/logical/ConvertCountToDirectScanRule.java
… exec/java-exec/src/main/java/org/apache/drill/exec/planner/logical/ConvertMetadataAggregateToDirectScanRule.java
… exec/java-exec/src/main/java/org/apache/drill/exec/planner/physical/DrillDistributionTrait.java
… exec/java-exec/src/main/java/org/apache/drill/exec/planner/physical/HashAggPrule.java
… exec/java-exec/src/main/java/org/apache/drill/exec/planner/physical/HashPrelUtil.java
… exec/java-exec/src/main/java/org/apache/drill/exec/planner/physical/MetadataAggPrel.java
… exec/java-exec/src/main/java/org/apache/drill/exec/planner/physical/MetadataAggPrule.java