Jira - [DRILL-7450](https://issues.apache.org/jira/browse/DRILL-7450).

This pull request contains several commits that may be reviewed one-by-one.
The first commit "Improve performance for `ANALYZE` command" contains changes for two-phase metadata stream aggregation support. The main logic is placed in `MetadataAggPrule` - it converts the lowest `MetadataAggRel` into the following plan:

```
MetadataAggPrel(PHASE_2of2)
  \
  HashToMergeExchangePrel
    \
    MetadataAggPrel(PHASE_1of2)
```

Also, this commit introduces `ConvertMetadataAggregateToDirectScanRule` rule which replaces the lowest `MetadataAggRel` with `DrillDirectScanRel` populated from row groups metadata if required metadata is present:

```
MetadataAggRel(metadataLevel=ROW_GROUP)
  \
  DrillScanRel
```

is converted into

```
DrillDirectScanRel
```

The last commit contains changes to use hash aggregation for `PHASE_1of2` for `ANALYZE` to reduce memory usage and avoid sorting non-aggregated data.
Performance measurements for ANALYZE command with and without these changes (Drill is running in embedded mode, 4 cores CPU, used tpch lineitem table):

| data table size | without changes | with changes |
| --- | --- | --- |
| 53 MB | 29.497 s | 26.871 s (1 hash agg fragment)
| 106.1 MB | 38.979 s | 21.381 s (3 hash agg fragments)
212.6 MB  | 57.846 s (external sort spilled data) | 26.629 s (3 hash agg fragments)
298.4 MB  | 59.42 s (external sort spilled data) | 25.482 s (3 hash agg fragments)
596.9 MB  | 142.14 s external sort spilled data) | 52.552 s (3 hash agg fragments)
100 GB (4-nodes cluster) | Failed with OOM after 45 minutes | 7.45 minutes